\name{PenIC-package}
\alias{PenIC-package}
\docType{package}
\title{
A statistical package for regression analysis of interval-censored data under the generalized odds-rates model using penalized B-splines
}
\description{
This package is designed to conduct the semiparametric regression analysis of interval-censored data under the generalized odds-rates model. To estimate the unknown nondecreasing cumulative baseline hazard function, monotone B-splines are used. An expectation maximization (EM) algorithm is developed to facilitate model fitting.
}
\details{
\tabular{ll}{
Package: \tab PenIC\cr
Type: \tab Package\cr
Version: \tab 1.0.0\cr
Date: \tab 2019-12-11\cr
}
}

\author{  Yan Liu  and Minggen Lu

}
